+++ login-workflow/example/src/screens/RegistrationProviderExample.tsx
@@ -14,7 +14,6 @@ import {
     // WorkflowCardHeader,
     // WorkflowCardBody,
     // WorkflowCardActions,
-    ChangePasswordDialogBase,
 } from '@brightlayer-ui/react-native-auth-workflow';
 import { RootStackParamList } from '../navigation';
 import { useNavigation } from '@react-navigation/native';
