--- login-workflow/src/components/ChangePasswordDialog/index.ts
@@ -1,2 +1,0 @@
-export * from './ChangePasswordDialogBase';
-export * from './types';
