+++ login-workflow/src/components/index.tsx
@@ -3,5 +3,4 @@ export * from './SetPassword';
 export * from './Error';
 export * from './Dialog';
 export * from './RegistrationWorkflow';
-export * from './ChangePasswordDialog';
 export * from './Spinner';
