+++ login-workflow/src/screens/ChangePasswordScreen/index.ts
@@ -1,0 +1,2 @@
+export * from './ChangePasswordScreenBase';
+export * from './types';
